feat(meetings): extract the agenda field with templates and AI helper
The agenda textarea and its two header popovers — Templates and Help me
write — lived inside the Agenda & Resources section, which is a drawer
section the quick dialog never renders. Both surfaces need the same
field, so it moves to `ComposerAgendaFieldComponent` and the section
composes it. The section keeps documents and links and loses nothing
else; its spec follows the code it covers into the field's own spec.

The popovers drop the theme's arrow and append to `body` with the page
scroll blocked, so a panel stays on the button it hangs from instead of
pointing at its own edge or sliding away underneath. The AI panel loses
its character counter — the agenda below it already has one, and two
counters on one screen invite the reader to work out which is which.

Documents showed two nested cards, because `lfx-file-upload` draws
PrimeNG's own advanced-mode chrome around the drop zone its consumers
project. The wrapper now hides that chrome, which is what every one of
its three consumers already assumed.

Refs #1458
