+++ apps/lfx-one/src/app/modules/meetings/meeting-composer/sections/composer-agenda-resources.component.html
@@ -7,92 +7,7 @@
     <p class="text-sm text-gray-500">Share agenda, documents, and reference materials so participants come prepared.</p>
   </div>
 
-  <div class="flex flex-col gap-1.5">
-    <div class="flex items-center justify-between gap-3">
-      <label for="composer-agenda" class="text-sm font-medium text-gray-900">Agenda</label>
-      <div class="flex items-center gap-2">
-        <lfx-button
-          label="Templates"
-          icon="fa-light fa-list"
-          size="small"
-          [text]="true"
-          [ariaPressed]="showTemplates()"
-          (onClick)="onToggleTemplates($event, templatesPopover)"
-          data-testid="composer-agenda-templates-toggle" />
-        <lfx-button
-          label="Help me write"
-          icon="fa-light fa-wand-magic-sparkles"
-          size="small"
-          [text]="true"
-          [ariaPressed]="showAiHelper()"
-          (onClick)="aiPopover.toggle($event)"
-          data-testid="composer-agenda-ai-toggle" />
-      </div>
-    </div>
-
-    <!-- Both header actions open popovers anchored to their button, per the design. `p-popover` is used
-         raw because the design system has no wrapper for it yet. -->
-    <p-popover
-      #templatesPopover
-      appendTo="body"
-      styleClass="w-[26rem] max-w-[calc(100vw-2rem)]"
-      (onShow)="showTemplates.set(true)"
-      (onHide)="showTemplates.set(false)">
-      @if (meetingType(); as type) {
-        <lfx-agenda-template-selector [meetingType]="type" (templateSelected)="onApplyTemplate($event, templatesPopover)" />
-      }
-    </p-popover>
-
-    <p-popover #aiPopover appendTo="body" styleClass="w-[26rem] max-w-[calc(100vw-2rem)]" (onShow)="showAiHelper.set(true)" (onHide)="onAiHelperHide()">
-      <div class="flex flex-col gap-3" data-testid="composer-agenda-ai-panel">
-        <div class="flex items-start gap-2.5">
-          <span class="flex h-7 w-7 shrink-0 items-center justify-center rounded-full bg-blue-600 text-xs text-white" aria-hidden="true">
-            <i class="fa-light fa-lightbulb"></i>
-          </span>
-          <div class="flex flex-col gap-0.5">
-            <h3 class="text-sm font-semibold text-gray-900">AI agenda generator</h3>
-            <p class="text-xs text-gray-500">Tell me the goal and I'll draft a structured agenda.</p>
-          </div>
-        </div>
-
-        <lfx-textarea
-          [form]="form()"
-          control="aiPrompt"
-          id="composer-agenda-ai-prompt"
-          placeholder="e.g. Review Q3 roadmap progress and agree on the release date"
-          styleClass="w-full"
-          [rows]="3"
-          [maxlength]="promptMaxLength"
-          data-testid="composer-agenda-ai-prompt" />
-        <p class="self-end text-xs" [ngClass]="aiPromptCounterClass()" data-testid="composer-agenda-ai-prompt-counter">
-          {{ aiPromptLength() }}/{{ promptMaxLength }}
-        </p>
-
-        <div class="flex items-center justify-end gap-2">
-          <lfx-button label="Cancel" size="small" severity="secondary" [text]="true" (onClick)="aiPopover.hide()" data-testid="composer-agenda-ai-cancel" />
-          <lfx-button
-            [label]="isGeneratingAgenda() ? 'Generating…' : 'Generate agenda'"
-            size="small"
-            severity="info"
-            [loading]="isGeneratingAgenda()"
-            [disabled]="isGeneratingAgenda()"
-            (onClick)="onGenerateAgenda(aiPopover)"
-            data-testid="composer-agenda-ai-generate" />
-        </div>
-      </div>
-    </p-popover>
-
-    <lfx-textarea
-      [form]="form()"
-      control="description"
-      id="composer-agenda"
-      placeholder="Enter meeting agenda and key discussion points"
-      styleClass="w-full min-h-48"
-      [rows]="8"
-      [maxlength]="agendaMaxLength"
-      data-testid="composer-agenda-textarea" />
-    <p class="self-end text-xs" [ngClass]="agendaCounterClass()">{{ agendaLength() }}/{{ agendaMaxLength }}</p>
-  </div>
+  <lfx-composer-agenda-field [form]="form()" />
 
   <div class="border-t border-gray-200"></div>
